(PC-36209) refactor(home): queryfy useGetVenuesData
